feat: Adding lazy loading
