+++ public/index.html
@@ -28,8 +28,8 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <!-- <div id="root"></div> -->
-    <my-counter name="rizwan Khan" />
+    <div id="root"></div>
+    <!-- <my-counter name="Rizwan Khan" /> -->
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
