+++ src/index.js
@@ -1,45 +1,15 @@
-/* eslint-disable react/no-typos */
-// This file is a web-component.
-import React from "react";
+import React, { Suspense, lazy } from "react";
 import ReactDOM from "react-dom";
-import reactToWebComponent from "react-to-webcomponent";
-import PropTypes from "prop-types";
 
-class MyCounter extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      count: 0,
-    };
-  }
-  inc() {
-    this.setState((prev) => ({
-      count: prev + 1,
-    }));
-  }
-  dec() {
-    this.setState((prev) => ({
-      count: prev - 1,
-    }));
-  }
-  render() {
-    return (
-      <div>
-        <h1>Counter {this.props.name} </h1>
-        <button onClick={() => this.inc()}>Up</button>
-        <button onClick={() => this.dec()}>Down</button>
-        <span>{this.state.count}</span>
-      </div>
-    );
-  }
+const MyCounter = lazy(() => import("./react-component"));
+
+function MyComponent() {
+  return (
+    <div>
+      <Suspense fallback={<div>Loading...</div>}>
+        <MyCounter name="Rizwan Khan" />
+      </Suspense>
+    </div>
+  );
 }
-
-MyCounter.PropTypes = {
-  name: PropTypes.string,
-};
-
-customElements.define(
-  "my-counter",
-  reactToWebComponent(MyCounter, React, ReactDOM)
-);
-// ReactDOM.render(<MyCounter />, document.getElementById("root"));
+ReactDOM.render(<MyComponent />, document.getElementById("root"));
